requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
frozenlist==1.2.0
idna==3.3
idna-ssl==1.1.0
multidict==5.2.0
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.2.post1
typing_extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
